refactor geecs_paths to scan_paths + utils
